Pydetect/requirements.txt
python-zbase32
